NSD online works, no google sign in

+++ app/src/main/java/com/game/team9/slidingpuzzle/MainActivity.java
@@ -1,49 +1,42 @@
 package com.game.team9.slidingpuzzle;
 
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
-
 import com.game.team9.slidingpuzzle.database.HighScoreDatabase;
 import com.game.team9.slidingpuzzle.network.MathModeService;
+import com.game.team9.slidingpuzzle.network.WifiService;
 
 public class MainActivity extends AppCompatActivity {
+
+
+    private static final Object m_Lock = new Object();
+    private static boolean m_Registered;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        Thread.setDefaultUncaughtExceptionHandler((t, e) -> {
-            AlertDialog.Builder alert = new AlertDialog.Builder(AppController.getInstance());
-            alert.setMessage(e.getMessage()).setNeutralButton("OK",
-                    new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            System.exit(0);
-                            android.os.Process.killProcess(android.os.Process.myPid());
-                        }
-                    });
-            alert.create().show();
-        });
-
+        setSupportActionBar(findViewById(R.id.toolbar));
         Thread.currentThread().setName("Main UI");
-        HighScoreDatabase.Initialize(getApplicationContext());
-        Intent intent = new Intent(this, MathModeService.class);
-        startService(intent);
+        synchronized (m_Lock) {
+            if(!m_Registered) {
+                HighScoreDatabase.Initialize(getApplicationContext());
+                startService(new Intent(this, MathModeService.class));
+              //  startService(new Intent(this, WifiService.class));
+                m_Registered = true;
+            }
+        }
     }
 
     public void numberOnClick(View view) throws Exception {
-        Intent x = new Intent(MainActivity.this, NumberModeMenuActivity.class);
-        startActivity(x);
+        startActivity(new Intent(MainActivity.this, NumberModeMenuActivity.class));
     }
 
     public void mathOnClick(View view)
     {
-        Intent y = new Intent(MainActivity.this, MathNameActivity.class);
-        startActivity(y);
+        startActivity(new Intent(MainActivity.this, MathNameActivity.class));
     }
 
 
@@ -52,9 +45,14 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        Intent intent = new Intent(this, MathModeService.class);
-        stopService(intent);
-        Log.i("Main", "Killing database");
-        HighScoreDatabase.DestroyInstance();
+        synchronized (m_Lock) {
+            if(m_Registered) {
+                stopService(new Intent(this, MathModeService.class));
+              //  stopService(new Intent(this, WifiService.class));
+                HighScoreDatabase.DestroyInstance();
+                Log.i("Main", "Killing database");
+                m_Registered = false;
+            }
+        }
     }
 }
